Re-added HPM support for configuration settings
Plugin settings should be relative to the the libconfig file root. For
example, a configuration setting of type HPCT_CHAR will be relative to
the root of conf/char/char-server.conf. In order to add a configuration
entry inside the char_configuration block, the full configuration path
(slash-delimited) should be passed to addCharConf(), as in the
following example:

`addCharConf("char_configuration/my_setting", my_parser_function);`

Signed-off-by: Haru <haru@dotalux.com>